feat(src/): rename component and practice 'reactive'
